refactor: organize retry tests by retry strategy
